+++ example/example.js
@@ -1,27 +1,12 @@
 /* eslint-disable no-console */
-const nodepub = require('../src/index.js');
+import { createEpub } from '../lib/index.js';
 
-// Metadata example.
-const metadata = {
-  id: '278-123456789',
-  cover: 'test/test-cover.png',
-  title: 'Unnamed Document',
-  series: 'My Series',
-  sequence: 1,
-  author: 'Anonymous',
-  fileAs: 'Anonymous',
-  genre: 'Non-Fiction',
-  tags: 'Sample,Example,Test',
-  copyright: 'Anonymous, 1980',
-  publisher: 'My Fake Publisher',
-  published: '2000-12-31',
-  language: 'en',
-  description: 'A test book.',
-  contents: 'Chapters',
-  showContents: true,
-  source: 'http://www.kcartlidge.com',
-  images: ['test/hat.png'],
-};
+const css = `body { font-family:Verdana,Arial,Sans-Serif; font-size:11pt; }
+#title,#title h1,#title h2,#title h3 { text-align:center; }
+h1,h3,p { margin-bottom:1em; }
+h2 { margin-bottom:2em; }
+p { text-indent: 0; }
+p+p { text-indent: 0.75em; }`;
 
 const copyright = `<h1>[[TITLE]]</h1>
 <h2>[[AUTHOR]]</h2>
@@ -69,7 +54,8 @@ const about = `<h1>About the Author</h1>
 `;
 
 // Dummy text (lorem ipsum).
-let lipsum = '<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse mattis iaculis pharetra. Proin malesuada tortor ut nibh viverra eleifend.</p><p>Duis efficitur, arcu vitae viverra consectetur, nisi mi pharetra metus, vel egestas ex velit id leo. Curabitur non tortor nisi. Mauris ornare, tellus vel fermentum suscipit, ligula est eleifend dui, in elementum nunc risus in ipsum. Pellentesque finibus aliquet turpis sed scelerisque. Pellentesque gravida semper elit, ut consequat est mollis sit amet. Nulla facilisi.</p>';
+let lipsum =
+  '<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse mattis iaculis pharetra. Proin malesuada tortor ut nibh viverra eleifend.</p><p>Duis efficitur, arcu vitae viverra consectetur, nisi mi pharetra metus, vel egestas ex velit id leo. Curabitur non tortor nisi. Mauris ornare, tellus vel fermentum suscipit, ligula est eleifend dui, in elementum nunc risus in ipsum. Pellentesque finibus aliquet turpis sed scelerisque. Pellentesque gravida semper elit, ut consequat est mollis sit amet. Nulla facilisi.</p>';
 for (let i = 0; i < 3; i += 1) {
   lipsum += lipsum;
 }
@@ -90,43 +76,80 @@ const generateContentsPage = (links) => {
   return contents;
 };
 
+const sections = [
+  {
+    content:
+      "<div id='title'><h1>[[TITLE]]</h1><h2>Book <strong>[[SEQUENCE]]</strong> of <em>[[SERIES]]</em></h2><h3>[[AUTHOR]]</h3><p> &nbsp;</p><p>&copy; [[COPYRIGHT]]</p></div>",
+    excludeFromContents: true,
+    isFrontMatter: true,
+    title: 'Title Page',
+  },
+  {
+    content: copyright,
+    filename: 'copyright-page',
+    isFrontMatter: true,
+    title: 'Copyright',
+  },
+  {
+    content: `<h1>One</h1>${lipsum}<p><a href='s3.xhtml'>A test internal link</a>.</p>`,
+    title: 'Chapter 1',
+  },
+  {
+    content: `<h1>Two</h1>${lipsum}`,
+    title: 'Chapter 2',
+  },
+  {
+    content: `<h1>Two (A)</h1><p><strong>This chapter does not appear in the contents.</strong></p>${lipsum}`,
+    excludeFromContents: true,
+    title: 'Chapter 2a',
+  },
+  {
+    content: `<h1>Three</h1><p>Here is a sample list.</p><ul><li>Sample list item one.</li><li>Sample list item two.</li><li>Sample list item three.</li></ul>${lipsum}`,
+    title: 'Chapter 3',
+  },
+  {
+    content: more,
+    title: 'More Books to Read',
+  },
+  {
+    content: about,
+    title: 'About the Author',
+  },
+];
+
+// Metadata example.
+const metadata = {
+  author: 'Anonymous',
+  contents: 'Chapters',
+  copyright: 'Anonymous, 1980',
+  cover: 'test/test-cover.png',
+  coverType: 'image',
+  css,
+  description: 'A test book.',
+  fileAs: 'Anonymous',
+  genre: 'Non-Fiction',
+  id: '278-123456789',
+  images: ['test/hat.png'],
+  language: 'en',
+  published: '2000-12-31',
+  publisher: 'My Fake Publisher',
+  sections,
+  sequence: 1,
+  series: 'My Series',
+  showContents: true,
+  source: 'http://www.kcartlidge.com',
+  tags: ['Sample', 'Example', 'Test'],
+  title: 'Unnamed Document',
+};
+
 // Set up the EPUB basics.
-const epub = nodepub.document(metadata, generateContentsPage);
-epub.addCSS(`body { font-family:Verdana,Arial,Sans-Serif; font-size:11pt; }
-#title,#title h1,#title h2,#title h3 { text-align:center; }
-h1,h3,p { margin-bottom:1em; }
-h2 { margin-bottom:2em; }
-p { text-indent: 0; }
-p+p { text-indent: 0.75em; }`);
-
-// Add some front matter.
-epub.addSection('Title Page', "<div id='title'><h1>[[TITLE]]</h1><h2>Book <strong>[[SEQUENCE]]</strong> of <em>[[SERIES]]</em></h2><h3>[[AUTHOR]]</h3><p> &nbsp;</p><p>&copy; [[COPYRIGHT]]</p></div>", true, true);
-epub.addSection('Copyright', copyright, false, true, 'copyright-page');
-
-// Add some content.
-epub.addSection('Chapter 1', `<h1>One</h1>${lipsum}<p><a href='chapter2.xhtml'>A test internal link</a>.</p>`);
-epub.addSection('Chapter 2', `<h1>Two</h1>${lipsum}`, false, false, 'chapter2');
-epub.addSection('Chapter 2a', `<h1>Two (A)</h1><p><strong>This chapter does not appear in the contents.</strong></p>${lipsum}`, true);
-epub.addSection('Chapter 3', `<h1>Three</h1><p>Here is a sample list.</p><ul><li>Sample list item one.</li><li>Sample list item two.</li><li>Sample list item three.</li></ul>${lipsum}`);
-epub.addSection('More Books to Read', more);
-epub.addSection('About the Author', about);
+const epub = createEpub(metadata, generateContentsPage);
 
 // Generate the result.
 (async () => {
   try {
     console.log('Generating a stand-alone EPUB.');
-    await epub.writeEPUB('example', 'example');
-  } catch (e) {
-    console.log('ERROR');
-    console.log(e);
-  }
-})();
-
-// Also write the structure both for debugging purposes and also to provide sample output in GitHub.
-(async () => {
-  try {
-    console.log('Generating a collection of EPUB constituent files.');
-    await epub.writeFilesForEPUB('example/example-EPUB-files');
+    await epub.write('example', 'example');
   } catch (e) {
     console.log('ERROR');
     console.log(e);